# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ProviEmplea",
        "description": "API REST para la gestión de talentos en ProviEmplea",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/empresas": {
            "get": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Listar empresas validadas",
                "operationId": "getEmpresas",
                "parameters": [
                    {
                        "name": "tipo_empresa",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "contratacion-directa",
                                "est",
                                "outsourcing"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Empresa"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Registrar nueva empresa",
                "operationId": "createEmpresa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmpresaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empresa creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Empresa"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/empresas/{id}": {
            "get": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Obtener empresa por ID",
                "operationId": "getEmpresa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Empresa"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Actualizar empresa",
                "operationId": "updateEmpresa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmpresaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa actualizada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Desactivar empresa",
                "operationId": "deleteEmpresa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Desactivada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            }
        },
        "/empresas/{id}/validar": {
            "patch": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Validar empresa",
                "operationId": "validarEmpresa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa validada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            }
        },
        "/personas": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Listar personas",
                "operationId": "getPersonas",
                "parameters": [
                    {
                        "name": "validado",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "nivel_educacional",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PersonaCVCiego"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Personas"
                ],
                "summary": "Registrar persona",
                "operationId": "createPersona",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creada"
                    }
                }
            }
        },
        "/personas/{id}": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Obtener por ID",
                "operationId": "getPersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "Personas"
                ],
                "summary": "Actualizar",
                "operationId": "updatePersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Personas"
                ],
                "summary": "Desactivar",
                "operationId": "deletePersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/personas/{id}/validar": {
            "patch": {
                "tags": [
                    "Personas"
                ],
                "summary": "Validar",
                "operationId": "validarPersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Empresa": {
                "required": [
                    "id",
                    "nombre_empresa",
                    "rut_empresa"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "nombre_empresa": {
                        "type": "string"
                    },
                    "rut_empresa": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "tipo_empresa": {
                        "type": "string"
                    },
                    "activo": {
                        "type": "boolean"
                    },
                    "validado": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "EmpresaInput": {
                "required": [
                    "nombre_empresa",
                    "rut_empresa",
                    "email",
                    "tipo_empresa",
                    "contacto_nombre",
                    "contacto_email"
                ],
                "properties": {
                    "nombre_empresa": {
                        "type": "string"
                    },
                    "rut_empresa": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "tipo_empresa": {
                        "type": "string",
                        "enum": [
                            "contratacion-directa",
                            "est",
                            "outsourcing"
                        ]
                    },
                    "contacto_nombre": {
                        "type": "string"
                    },
                    "contacto_email": {
                        "type": "string",
                        "format": "email"
                    },
                    "contacto_telefono": {
                        "type": "string",
                        "nullable": true
                    },
                    "rubro": {
                        "type": "string",
                        "nullable": true
                    },
                    "presentacion": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "PersonaCVCiego": {
                "required": [
                    "id",
                    "nivel_educacional"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "nivel_educacional": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PersonaInput": {
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "apellido": {
                        "type": "string"
                    },
                    "nivel_educacional": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
...
}